Removed deprecated, more inits and methods for TransactionsOptions and EthereumTransaction that were in Legacy, fixed assemblePromise for WriteTransaction
