+++ web3swift/PrecompiledContracts/ERC1400/Web3+ERC1400.swift
@@ -66,9 +66,6 @@ protocol IERC1400: IERC20 {
 // can be imperatively read and saved
 public class ERC1400: IERC1400 {
     
-    @available(*, deprecated, renamed: "transactionOptions")
-    public var options: Web3Options = .init()
-    
     private var _name: String? = nil
     private var _symbol: String? = nil
     private var _decimals: UInt8? = nil
