+++ web3swift/PrecompiledContracts/ERC1594/Web3+ERC1594.swift
@@ -34,9 +34,6 @@ protocol IERC1594: IERC20 {
 
 public class ERC1594: IERC1594 {
     
-    @available(*, deprecated, renamed: "transactionOptions")
-    public var options: Web3Options = .init()
-    
     private var _name: String? = nil
     private var _symbol: String? = nil
     private var _decimals: UInt8? = nil
