+++ web3swift/PrecompiledContracts/ERC1633/Web3+ERC1633.swift
@@ -21,9 +21,6 @@ protocol IERC1633: IERC20, IERC165 {
 
 public class ERC1633: IERC1633 {
     
-    @available(*, deprecated, renamed: "transactionOptions")
-    public var options: Web3Options = .init()
-    
     private var _name: String? = nil
     private var _symbol: String? = nil
     private var _decimals: UInt8? = nil
