+++ web3swift/PrecompiledContracts/ERC888/Web3+ERC888.swift
@@ -19,9 +19,6 @@ protocol IERC888 {
 
 public class ERC888: IERC888 {
     
-    @available(*, deprecated, renamed: "transactionOptions")
-    public var options: Web3Options = .init()
-    
     private var _name: String? = nil
     private var _symbol: String? = nil
     private var _decimals: UInt8? = nil
