+++ web3swift/Web3/Classes/Web3+Options.swift
@@ -117,66 +117,70 @@ public struct TransactionOptions {
         return opts
     }
 
-//    public static func fromJSON(_ json: [String: Any]) -> Web3Options? {
-//        var options = Web3Options()
-//        if let gas = json["gas"] as? String, let gasBiguint = BigUInt(gas.stripHexPrefix().lowercased(), radix: 16) {
-//            options.gasLimit = gasBiguint
-//        }
-//        if let gasPrice = json["gasPrice"] as? String, let gasPriceBiguint = BigUInt(gasPrice.stripHexPrefix().lowercased(), radix: 16) {
-//            options.gasPrice = gasPriceBiguint
-//        }
-//        if let value = json["value"] as? String, let valueBiguint = BigUInt(value.stripHexPrefix().lowercased(), radix: 16) {
-//            options.value = valueBiguint
-//        }
-//        if let fromString = json["from"] as? String {
-//            guard let addressFrom = EthereumAddress(fromString) else {return nil}
-//            options.from = addressFrom
-//        }
-//        return options
-//    }
-//
-//    /// Merges two sets of topions by overriding the parameters from the first set by parameters from the second
-//    /// set if those are not nil.
-//    ///
-//    /// Returns default options if both parameters are nil.
-//    public static func merge(_ options:Web3Options?, with other:Web3Options?) -> Web3Options? {
-//        if (other == nil && options == nil) {
-//            return Web3Options.defaultOptions()
-//        }
-//        var newOptions = Web3Options.defaultOptions()
-//        if (other?.to != nil) {
-//            newOptions.to = other?.to
-//        } else {
-//            newOptions.to = options?.to
-//        }
-//        if (other?.from != nil) {
-//            newOptions.from = other?.from
-//        } else {
-//            newOptions.from = options?.from
-//        }
-//        if (other?.gasLimit != nil) {
-//            newOptions.gasLimit = other?.gasLimit
-//        } else {
-//            newOptions.gasLimit = options?.gasLimit
-//        }
-//        if (other?.gasPrice != nil) {
-//            newOptions.gasPrice = other?.gasPrice
-//        } else {
-//            newOptions.gasPrice = options?.gasPrice
-//        }
-//        if (other?.value != nil) {
-//            newOptions.value = other?.value
-//        } else {
-//            newOptions.value = options?.value
-//        }
-//        return newOptions
-//    }
-//
+    public static func fromJSON(_ json: [String: Any]) -> TransactionOptions? {
+        var options = TransactionOptions()
+        if let gas = json["gas"] as? String, let gasBiguint = BigUInt(gas.stripHexPrefix().lowercased(), radix: 16) {
+            options.gasLimit = .manual(gasBiguint)
+        } else {
+            options.gasLimit = .automatic
+        }
+        if let gasPrice = json["gasPrice"] as? String, let gasPriceBiguint = BigUInt(gasPrice.stripHexPrefix().lowercased(), radix: 16) {
+            options.gasPrice = .manual(gasPriceBiguint)
+        } else {
+            options.gasPrice = .automatic
+        }
+        if let value = json["value"] as? String, let valueBiguint = BigUInt(value.stripHexPrefix().lowercased(), radix: 16) {
+            options.value = valueBiguint
+        }
+        if let fromString = json["from"] as? String {
+            guard let addressFrom = EthereumAddress(fromString) else {return nil}
+            options.from = addressFrom
+        }
+        return options
+    }
+
+    /// Merges two sets of topions by overriding the parameters from the first set by parameters from the second
+    /// set if those are not nil.
+    ///
+    /// Returns default options if both parameters are nil.
+    public static func merge(_ options:TransactionOptions?, with other:TransactionOptions?) -> TransactionOptions? {
+        if (other == nil && options == nil) {
+            return TransactionOptions.defaultOptions
+        }
+        var newOptions = TransactionOptions.defaultOptions
+        if (other?.to != nil) {
+            newOptions.to = other?.to
+        } else {
+            newOptions.to = options?.to
+        }
+        if (other?.from != nil) {
+            newOptions.from = other?.from
+        } else {
+            newOptions.from = options?.from
+        }
+        if (other?.gasLimit != nil) {
+            newOptions.gasLimit = other?.gasLimit
+        } else {
+            newOptions.gasLimit = options?.gasLimit
+        }
+        if (other?.gasPrice != nil) {
+            newOptions.gasPrice = other?.gasPrice
+        } else {
+            newOptions.gasPrice = options?.gasPrice
+        }
+        if (other?.value != nil) {
+            newOptions.value = other?.value
+        } else {
+            newOptions.value = options?.value
+        }
+        return newOptions
+    }
+
 //    /// merges two sets of options along with a gas estimate to try to guess the final gas limit value required by user.
 //    ///
 //    /// Please refer to the source code for a logic.
-//    public static func smartMergeGasLimit(originalOptions: Web3Options?, extraOptions: Web3Options?, gasEstimate: BigUInt) -> BigUInt? {
-//        guard let mergedOptions = Web3Options.merge(originalOptions, with: extraOptions) else {return nil} //just require any non-nils
+//    public static func smartMergeGasLimit(originalOptions: TransactionOptions?, extraOptions: TransactionOptions?, gasEstimate: BigUInt) -> BigUInt? {
+//        guard let mergedOptions = TransactionOptions.merge(originalOptions, with: extraOptions) else {return nil} //just require any non-nils
 //        if mergedOptions.gasLimit == nil {
 //            return gasEstimate // for user's convenience we just use an estimate
 //            //            return nil // there is no opinion from user, so we can not proceed
@@ -198,7 +202,7 @@ public struct TransactionOptions {
 //            }
 //        }
 //    }
-//
+
 //    public static func smartMergeGasPrice(originalOptions: Web3Options?, extraOptions: Web3Options?, priceEstimate: BigUInt) -> BigUInt? {
 //        guard let mergedOptions = Web3Options.merge(originalOptions, with: extraOptions) else {return nil} //just require any non-nils
 //        if mergedOptions.gasPrice == nil {
