--- web3swift/web3swiftLegacy/ABIv2Legacy/Classes/ABIv2Legacy.swift
@@ -1,14 +1,0 @@
-//
-//  ABIv2.swift
-//  web3swift
-//
-//  Created by Alexander Vlasov on 02.04.2018.
-//  Copyright © 2018 Bankex Foundation. All rights reserved.
-//
-
-import Foundation
-
-@available(*, deprecated, message: "Use ABI pod instead")
-public struct ABIv2 {
-    
-}
