--- web3swift/web3swiftLegacy/ContractLegacy/Classes/ContractABIv2Legacy.swift
@@ -1,17 +1,0 @@
-//
-//  ContractABIv2.swift
-//  web3swift
-//
-//  Created by Alexander Vlasov on 04.04.2018.
-//  Copyright © 2018 Bankex Foundation. All rights reserved.
-//
-
-import Foundation
-import BigInt
-import EthereumAddress
-import EthereumABI
-
-@available(*, deprecated, message: "Use EthereumContract instead")
-public struct ContractV2 {
-    
-}
