--- web3swift/web3swiftLegacy/KeystoreManagerLegacy/Classes/AbstractKeystoreLegacy.swift
@@ -1,13 +1,0 @@
-//
-//  AbstractKeystore.swift
-//  web3swift
-//
-//  Created by Alexander Vlasov on 10.01.2018.
-//  Copyright © 2018 Bankex Foundation. All rights reserved.
-//
-
-import Foundation
-
-@available(*, deprecated, message: "Use ReadTransaction or WriteTransaction instead")
-public typealias TransactionIntermediate = web3.web3contract.TransactionIntermediate
-
